Move RustCardanoPrimitives into lib
