+++ src/Transaction/Transaction.spec.ts
@@ -1,8 +1,8 @@
 import { expect } from 'chai'
-import { RustCardano } from '../Cardano'
 import Transaction, { TransactionInput, TransactionOutput } from './'
 import { InsufficientTransactionInput } from './errors'
 import { EmptyArray } from '../lib/validator/errors'
+import { RustCardano } from '../lib/RustCardanoPrimitives'
 
 describe('Transaction', () => {
   it('throws if inputs are invalid', () => {
