+++ src/Wallet/lib/address_derivation.spec.ts
@@ -5,7 +5,7 @@ import { SCAN_GAP } from '../config'
 import { generateTestTransaction, mockProvider, seedTransactionSet, generateTestUtxos } from '../../test/utils'
 import { deriveAddressSet } from '.'
 import { AddressType } from '..'
-import { RustCardano } from '../../Cardano'
+import { RustCardano } from '../../lib/RustCardanoPrimitives'
 
 describe('deriveAddressSet', () => {
   it('combines external and internal addresses up to the end of each range', async () => {
