+++ src/index.ts
@@ -3,7 +3,8 @@ import Wallet from './Wallet'
 import { Provider, HttpProvider } from './Provider'
 import * as Utils from './Utils'
 import { InMemoryKeyManager, LedgerKeyManager } from './KeyManager'
-import { RustCardano, FeeAlgorithm, ChainSettings } from './Cardano'
+import { FeeAlgorithm, ChainSettings } from './Cardano'
+import { RustCardano } from './lib/RustCardanoPrimitives'
 
 export const providers: { [provider: string]: (connection: string) => Provider } = {
   http: HttpProvider
