+++ src/test/utils/utxo.ts
@@ -1,7 +1,7 @@
 import { AddressType } from '../../Wallet'
 import { hexGenerator } from '.'
 import { addressDiscoveryWithinBounds } from '../../Utils'
-import { RustCardano } from '../../Cardano'
+import { RustCardano } from '../../lib/RustCardanoPrimitives'
 
 export function generateTestUtxos ({ account, lowerBound, upperBound, type, value }: { account: string, lowerBound: number, upperBound: number, type: AddressType, value: string }) {
   const numberOfUtxos = upperBound - lowerBound
